pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netscope"
version = "1.1.0"
description = "SSH-based network topology intelligence for Cisco networks"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0,<1.0",
    "uvicorn[standard]>=0.29.0,<1.0",
    "scrapli[asyncssh]>=2024.1.30,<2026.0",
    "asyncssh>=2.14.0,<3.0",
    "cryptography>=42.0.0,<45.0",
    "pydantic>=2.7.0,<3.0",
    "pydantic-settings>=2.2.0,<3.0",
    "openpyxl>=3.1.2,<4.0",
    "python-multipart>=0.0.9,<1.0",
    "aiofiles>=23.2.1,<25.0",
    "httpx>=0.27.0,<1.0",
    "pyyaml>=6.0.1,<7.0",
    "slowapi>=0.1.9,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0,<9.0",
    "pytest-asyncio>=0.23.6,<1.0",
    "httpx>=0.27.0,<1.0",
    "ruff>=0.4.4,<1.0",
    "mypy>=1.10.0,<2.0",
    "types-PyYAML>=6.0.12,<7.0",
    "pytest-cov>=5.0.0,<6.0",
    "pip-audit>=2.7.0,<3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=backend --cov-report=term-missing"
